+++ src/ui/help.rs
@@ -92,5 +92,6 @@ pub fn get_help_docs() -> Vec<Vec<&'static str>> {
     vec!["Delete saved album", "D", "Library -> Albums"],
     vec!["Delete saved playist", "D", "Playlist"],
     vec!["Follow an artists/playlist", "w", "Search result"],
+    vec!["Play random song in playlist", "S", "Selected Playlist"],
   ]
 }